Restructured the routes to be inside splash for session
